data_parsing/requirements.txt
requests
pydantic
python-dotenv
pandas
numpy
matplotlib
seaborn
tqdm

